Harmonize treatment of empty IANA considerations
Currently, when IANA performs its last-call review of documents
with no actions, the review text includes:

%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
We understand that this document doesn't require any registry actions.

While it's often helpful for a document's IANA Considerations section to
remain in place upon publication even if there are no actions, if the
authors strongly prefer to remove it, we do not object.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

This is not really consistent with a claim that the RFC Editor will
remove such sections.